+++ arch/x86_32/kernel/process.c
@@ -1,0 +1,6 @@
+#include <novium/cpu.h>
+#include <asm/cpu.h>
+
+void cpu_idle(void) {
+    cpu_hlt();
+}