+++ fs/fs.h
@@ -1,1 +1,1 @@
-// TODO: System stub - implement when ready.
+/* TODO: System stub - implement when ready. */