+++ include/drivers/input.h
@@ -3,7 +3,7 @@
 
 #include <novium/types.h>
 
-// keyboard - raw scancode interface.
+/* keyboard - raw scancode interface. */
 void keyboard_init(void);      
 int  keyboard_pop(void); 
 int keyboard_getchar(void);